+++ tests/default.nix
@@ -30,6 +30,7 @@ let
     // prefix "hostname" (tryImport "hostname.nix")
     // prefix "domain" (tryImport "domain.nix")
     // prefix "host" (tryImport "host.nix")
+    // prefix "vlanId" (tryImport "vlan-id.nix")
     // prefix "registry" (tryImport "registry.nix")
     // prefix "iparse" (tryImport "internal/parse.nix")
     // prefix "ifmt" (tryImport "internal/format.nix")